Rename motd to checkNewVersion, remove commented code
